Move Substitute instance of TermLike
